use the timestamp from the event to determine clicks
This avoids problems with a) clicks happening during other longer running coroutines and b) gradual loss of accuracy of `time.monotonic` over longer periods.

The implementation of keypad in Blinka does not use this field, so it will default to using the ticks library, and lose the benefit of a), but retain b). This is a requirement of the asyncio library anyway,